refactor(DAL): separate IAccountingService from implementation
